+++ CHANGELOG.md
@@ -137,8 +137,6 @@
         - Added
             - `unvalidated_redirect_dom` -- Logs DOM-based unvalidated redirects.
             - `x_frame_options` -- Logs missing `X-Frame-Options` headers per host.
-            - `insecure_cors_policy` -- Logs wildcard `Access-Control-Allow-Origin`
-                headers per host.
             - `xxe` -- Logs XML External Entity vulnerabilities.
         - `trainer` -- Disabled parameter flip for the payload to avoid parameter
             pollution.
@@ -164,6 +162,8 @@
                 files for wildcard `allow-http-request-headers-from` policies.
             - `insecure_client_access_policy` -- Checks `clientaccesspolicy.xml`
                 files for wildcard domain policies.
+            - `insecure_cors_policy` -- Logs wildcard `Access-Control-Allow-Origin`
+                headers per host.
             - `common_directories` -- Added:
                 - `rails/info/routes`
                 - `rails/info/properties`