+++ CHANGELOG.md
@@ -6,6 +6,9 @@
     - `#log_remote_file` -- Assign `HTTP::Response#status_line` as proof.
 - `Issue`
     - `#signature` -- Store `Regexp` source instead of converting it to String.
+- Checks
+    - Active
+        - `csrf` -- Assign form source as proof.
 
 ## _Under development_